Move support/historyarchive to top level (#2783)
Move `support/historyarchive` to top level.

According to monorepo README:

> support contains packages that are not intended for consumption
> outside of Stellar's other packages.

`historyarchive` is actually intended to use outside Stellar packages.
