+++ exp/ingest/adapters/history_archive_adapter.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/stellar/go/exp/ingest/io"
+	"github.com/stellar/go/historyarchive"
 	"github.com/stellar/go/support/errors"
-	"github.com/stellar/go/support/historyarchive"
 	"github.com/stellar/go/xdr"
 )
