+++ exp/ingest/ledgerbackend/captive_core_backend.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/stellar/go/support/historyarchive"
+	"github.com/stellar/go/historyarchive"
 	"github.com/stellar/go/support/log"
 	"github.com/stellar/go/xdr"
 )
