+++ exp/ingest/ledgerbackend/history_archive_backend.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"io"
 
+	"github.com/stellar/go/historyarchive"
 	"github.com/stellar/go/support/errors"
-	"github.com/stellar/go/support/historyarchive"
 	"github.com/stellar/go/xdr"
 )
