+++ historyarchive/hash.go
