+++ historyarchive/hash_test.go
