+++ historyarchive/verify.go
