+++ tools/archive-reader/archive_reader.go
@@ -8,7 +8,7 @@ import (
 	"log"
 
 	"github.com/stellar/go/exp/ingest/adapters"
-	"github.com/stellar/go/support/historyarchive"
+	"github.com/stellar/go/historyarchive"
 )
 
 func main() {
