+++ tools/stellar-archivist/main.go
@@ -12,8 +12,8 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/stellar/go/historyarchive"
 	"github.com/stellar/go/support/errors"
-	"github.com/stellar/go/support/historyarchive"
 )
 
 func status(a string, opts *Options) {
